refactor: park the operator consoles, keep their portable halves
cli/dev_envs.py was the control-plane worker's side of the materialization
protocol -- claim-materialization --worker-id, seed-topology-manifest -- and
research_factory_standup.py hardcoded factory.name == "synth-rsi", internal
milestone IDs, and Synth's private repo remotes. Neither is something a
customer can use, and both shipped in the wheel with console scripts.

Lifted before removing: the Cloud S0 evidence verifier becomes
verify_cloud_s0_evidence in sdk/research/session/dev_environments.py, and the
stand-up orchestration becomes typed research.factories.standup() /
plan_standup(), without the RSI activation gate that made the CLI internal.

Also out of the package: sdk/research/schemas/ moves to a top-level schemas/.
Nothing read those 2.1MB at runtime, and a Research-only wheel was shipping the
backend's /v1/tunnels and /v1/pools route definitions. They stay tracked because
testing/ compares them against the backend. factory_plans/ goes with the script
that read it -- its one plan is Synth's own kind:internal factory.

Docs now describe what the code does: the MCP README said tools ship under the
smr_ prefix (the registry renames every one to research_) and documented nine
tools the shipped entrypoint does not advertise; the session README linked into
one laptop's checkout of a deleted repo.

Wheel: 2.7MB -> 604K.
