+++ pyproject.toml
@@ -18,7 +18,6 @@ dependencies = [
 [project.scripts]
 synth-ai = "synth_ai.cli:cli"
 synth-ai-research-mcp = "synth_ai.mcp.research.server:main"
-synth-ai-research-factory-standup = "synth_ai.cli.research_factory_standup:main"
 
 [project.optional-dependencies]
 dev = [
@@ -66,12 +65,6 @@ include-package-data = true
 [tool.setuptools.package-data]
 "synth_ai" = [
     "py.typed",
-    # Research moved core/research -> sdk/research (see unify_sdk_layering.md).
-    # These globs are the only thing that puts the factory plans and vendored
-    # schemas into the wheel; a stale path here ships a package that imports and
-    # then fails at runtime looking for data files.
-    "sdk/research/factory_plans/*.json",
-    "sdk/research/schemas/*",
 ]
 
 [tool.ruff]
