+++ synth_ai/mcp/research/tools/visuals.py
@@ -1,7 +1,4 @@
-"""Stable MCP adapters for first-class Research Visuals.
-
-# See: Jstack/.jstack/daily_notes/2026-07-27/SPEC_visuals_resource.md
-"""
+"""Stable MCP adapters for first-class Research Visuals."""
 
 from __future__ import annotations
