--- synth_ai/sdk/research/factory_plans/__init__.py
@@ -1,1 +1,0 @@
-"""Packaged declarative Research Factory plans."""
