+++ synth_ai/sdk/research/session/dev_environments.py
@@ -856,4 +856,214 @@ class DevEnvironmentsAPI(_ClientNamespace):
         return ", ".join(pieces)
 
 
-__all__ = ["DevEnvironmentsAPI"]
+# The Cloud S0 receipt shape is a backend contract, so reading it lives next to
+# the namespace that fetches it.
+
+
+def _mapping_at(payload: dict[str, Any], *path: str) -> dict[str, Any]:
+    current: object = payload
+    for key in path:
+        if not isinstance(current, dict):
+            return {}
+        current = current.get(key)
+    return dict(current) if isinstance(current, dict) else {}
+
+
+def _list_at(payload: dict[str, Any], *path: str) -> list[object]:
+    current: object = payload
+    for key in path:
+        if not isinstance(current, dict):
+            return []
+        current = current.get(key)
+    return list(current) if isinstance(current, list) else []
+
+
+def _text_at(payload: dict[str, Any], *path: str) -> str | None:
+    current: object = payload
+    for key in path:
+        if not isinstance(current, dict):
+            return None
+        current = current.get(key)
+    text = str(current or "").strip()
+    return text or None
+
+
+def _positive_int(value: object) -> bool:
+    if isinstance(value, bool):
+        return False
+    if isinstance(value, int):
+        return value > 0
+    if isinstance(value, float):
+        return value > 0
+    try:
+        return int(str(value or "0")) > 0
+    except (TypeError, ValueError):
+        return False
+
+
+def _string_list(value: object) -> list[str]:
+    if not isinstance(value, list):
+        return []
+    return [text for item in value if (text := str(item or "").strip())]
+
+
+def _git_proofs_from_evidence(
+    payload: dict[str, Any],
+    *,
+    expected_run_ids: set[str],
+) -> list[dict[str, Any]]:
+    proofs: list[dict[str, Any]] = []
+    for item in _list_at(payload, "receipts", "items"):
+        if not isinstance(item, dict):
+            continue
+        if str(item.get("kind") or "") != "dev_environment_run_receipt":
+            continue
+        run_id = str(item.get("run_id") or "").strip()
+        if expected_run_ids and run_id not in expected_run_ids:
+            continue
+        git_raw = item.get("git")
+        git = dict(git_raw) if isinstance(git_raw, Mapping) else {}
+        run_git_raw = git.get("run_git_context")
+        run_git = dict(run_git_raw) if isinstance(run_git_raw, Mapping) else {}
+        project_git_raw = git.get("project_git")
+        project_git = dict(project_git_raw) if isinstance(project_git_raw, Mapping) else {}
+        branch = str(run_git.get("branch") or project_git.get("default_branch") or "").strip()
+        commit_sha = str(
+            run_git.get("head_commit_sha") or project_git.get("commit_sha") or ""
+        ).strip()
+        source_refs = (
+            dict(item.get("source_refs")) if isinstance(item.get("source_refs"), dict) else {}
+        )
+        if branch and commit_sha:
+            proofs.append(
+                {
+                    "run_id": run_id,
+                    "branch": branch,
+                    "commit_sha": commit_sha,
+                    "source": run_git.get("source") or "unknown",
+                    "last_push_confirmed": run_git.get("last_push_confirmed"),
+                    "project_git_status": source_refs.get("project_git_status"),
+                    "run_git": source_refs.get("run_git"),
+                }
+            )
+    return proofs
+
+
+def verify_cloud_s0_evidence(
+    payload: dict[str, Any],
+    *,
+    expected_dev_environment_id: str | None = None,
+    expected_project_id: str | None = None,
+    expected_run_ids: tuple[str, ...] = (),
+    expected_host_kind: str | None = "daytona",
+) -> dict[str, Any]:
+    summary = _mapping_at(payload, "summary")
+    run_binding_summary = _mapping_at(payload, "summary", "run_binding_summary")
+    cloud_s0_proof = _mapping_at(payload, "summary", "cloud_s0_proof")
+    cloud_s0_checks = _mapping_at(payload, "summary", "cloud_s0_proof", "checks")
+    environment = _mapping_at(payload, "environment")
+    bound_run_ids = _string_list(run_binding_summary.get("bound_run_ids"))
+    expected_run_id_set = {
+        run_id for item in expected_run_ids if (run_id := str(item or "").strip())
+    }
+    actual_dev_environment_id = (
+        _text_at(payload, "dev_environment_id")
+        or _text_at(summary, "dev_environment_id")
+        or _text_at(environment, "dev_environment_id")
+        or _text_at(environment, "environment_id")
+    )
+    actual_project_id = (
+        _text_at(summary, "project_id")
+        or _text_at(environment, "project_id")
+        or _text_at(payload, "project_id")
+    )
+    actual_host_kind = _text_at(summary, "host_kind") or _text_at(environment, "host_kind")
+    git_proofs = _git_proofs_from_evidence(
+        payload,
+        expected_run_ids=expected_run_id_set,
+    )
+
+    checks: dict[str, bool] = {}
+    missing: list[str] = []
+
+    def require(name: str, passed: bool) -> None:
+        checks[name] = bool(passed)
+        if not passed:
+            missing.append(name)
+
+    require("summary_present", bool(summary))
+    require("run_binding_summary_present", bool(run_binding_summary))
+    require("cloud_s0_proof_present", bool(cloud_s0_proof))
+    require("bound_run_ids_present", bool(bound_run_ids))
+    require("has_bound_run", run_binding_summary.get("has_bound_run") is True)
+    require(
+        "has_dev_slot_execution_binding",
+        run_binding_summary.get("has_dev_slot_execution_binding") is True,
+    )
+    require(
+        "has_daytona_binding",
+        run_binding_summary.get("has_daytona_binding") is True,
+    )
+    require("receipt_ready", cloud_s0_proof.get("receipt_ready") is True)
+    for key in (
+        "environment_id_bound",
+        "dev_slot_execution_bound",
+        "daytona_bound",
+        "work_product_present",
+        "raw_trace_present",
+        "cost_snapshot_present",
+        "usage_snapshot_present",
+    ):
+        require(f"cloud_s0_checks.{key}", cloud_s0_checks.get(key) is True)
+    require(
+        "work_product_count_positive",
+        _positive_int(cloud_s0_proof.get("work_product_count")),
+    )
+    require("trace_count_positive", _positive_int(cloud_s0_proof.get("trace_count")))
+    require(
+        "cost_summary_present",
+        bool(_mapping_at(payload, "summary", "cloud_s0_proof", "cost_summary")),
+    )
+    require(
+        "usage_summary_present",
+        bool(_mapping_at(payload, "summary", "cloud_s0_proof", "usage_summary")),
+    )
+    require("git_branch_and_sha_present", bool(git_proofs))
+    if expected_dev_environment_id:
+        require(
+            "expected_dev_environment_id",
+            actual_dev_environment_id == expected_dev_environment_id,
+        )
+    if expected_project_id:
+        require("expected_project_id", actual_project_id == expected_project_id)
+    if expected_host_kind:
+        require("expected_host_kind", actual_host_kind == expected_host_kind)
+    if expected_run_id_set:
+        require(
+            "expected_run_ids_present",
+            expected_run_id_set.issubset(set(bound_run_ids)),
+        )
+
+    return {
+        "ok": not missing,
+        "checks": checks,
+        "missing": missing,
+        "actual": {
+            "dev_environment_id": actual_dev_environment_id,
+            "project_id": actual_project_id,
+            "host_kind": actual_host_kind,
+            "bound_run_ids": bound_run_ids,
+            "git_proofs": git_proofs,
+            "work_product_count": cloud_s0_proof.get("work_product_count"),
+            "trace_count": cloud_s0_proof.get("trace_count"),
+        },
+        "expected": {
+            "dev_environment_id": expected_dev_environment_id,
+            "project_id": expected_project_id,
+            "host_kind": expected_host_kind,
+            "run_ids": sorted(expected_run_id_set),
+        },
+    }
+
+
+__all__ = ["DevEnvironmentsAPI", "verify_cloud_s0_evidence"]
